Rename 'activeactive' package to 'mcf' (#3615)
* Rename 'activeactive' package to 'mcf'

* Reformat imports in UnifiedJedis
