+++ src/main/java/redis/clients/jedis/mcf/CircuitBreakerCommandExecutor.java
@@ -1,33 +1,39 @@
-package redis.clients.jedis.activeactive;
+package redis.clients.jedis.mcf;
 
 import io.github.resilience4j.circuitbreaker.CircuitBreaker;
 import io.github.resilience4j.decorators.Decorators;
 import io.github.resilience4j.decorators.Decorators.DecorateSupplier;
 
+import redis.clients.jedis.CommandObject;
 import redis.clients.jedis.Connection;
+import redis.clients.jedis.executors.CommandExecutor;
 import redis.clients.jedis.providers.MultiClusterPooledConnectionProvider;
 import redis.clients.jedis.providers.MultiClusterPooledConnectionProvider.Cluster;
 
 /**
- * ConnectionProvider with built-in retry, circuit-breaker, and failover to another cluster/database endpoint.
+ * @author Allen Terleto (aterleto)
+ * <p>
+ * CommandExecutor with built-in retry, circuit-breaker, and failover to another cluster/database endpoint.
  * With this executor users can seamlessly failover to Disaster Recovery (DR), Backup, and Active-Active cluster(s)
  * by using simple configuration which is passed through from Resilience4j - https://resilience4j.readme.io/docs
+ * <p>
  */
-public class CircuitBreakerFailoverConnectionProvider extends CircuitBreakerFailoverBase {
+public class CircuitBreakerCommandExecutor extends CircuitBreakerFailoverBase implements CommandExecutor {
 
-    public CircuitBreakerFailoverConnectionProvider(MultiClusterPooledConnectionProvider provider) {
+    public CircuitBreakerCommandExecutor(MultiClusterPooledConnectionProvider provider) {
         super(provider);
     }
 
-    public Connection getConnection() {
+    @Override
+    public <T> T executeCommand(CommandObject<T> commandObject) {
         Cluster cluster = provider.getCluster(); // Pass this by reference for thread safety
 
-        DecorateSupplier<Connection> supplier = Decorators.ofSupplier(() -> this.handleGetConnection(cluster));
+        DecorateSupplier<T> supplier = Decorators.ofSupplier(() -> this.handleExecuteCommand(commandObject, cluster));
 
         supplier.withRetry(cluster.getRetry());
         supplier.withCircuitBreaker(cluster.getCircuitBreaker());
         supplier.withFallback(defaultCircuitBreakerFallbackException,
-                              e -> this.handleClusterFailover(cluster.getCircuitBreaker()));
+                              e -> this.handleClusterFailover(commandObject, cluster.getCircuitBreaker()));
 
         return supplier.decorate().get();
     }
@@ -35,21 +41,21 @@ public class CircuitBreakerFailoverConnectionProvider extends CircuitBreakerFail
     /**
      * Functional interface wrapped in retry and circuit breaker logic to handle happy path scenarios
      */
-    private Connection handleGetConnection(Cluster cluster) {
-        Connection connection = cluster.getConnection();
-        connection.ping();
-        return connection;
+    private <T> T handleExecuteCommand(CommandObject<T> commandObject, Cluster cluster) {
+        try (Connection connection = cluster.getConnection()) {
+            return connection.executeCommand(commandObject);
+        }
     }
 
     /**
      * Functional interface wrapped in retry and circuit breaker logic to handle open circuit breaker failure scenarios
      */
-    private Connection handleClusterFailover(CircuitBreaker circuitBreaker) {
+    private <T> T handleClusterFailover(CommandObject<T> commandObject, CircuitBreaker circuitBreaker) {
 
         clusterFailover(circuitBreaker);
 
         // Recursive call to the initiating method so the operation can be retried on the next cluster connection
-        return getConnection();
+        return executeCommand(commandObject);
     }
 
 }
