+++ src/main/java/redis/clients/jedis/mcf/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package contains the classes that are related to Active-Active cluster(s) and Multi-Cluster failover.
  */
-package redis.clients.jedis.activeactive;
+package redis.clients.jedis.mcf;
